bmco: Bayesian Analysis for Multivariate Categorical Outcomes

Provides Bayesian methods for comparing groups on multiple binary outcomes, including basic tests, regression adjustment, and multilevel models.

Main functions

bmvb: Bayesian test using multivariate Bernoulli

bglm: Subgroup analysis using Bayesian logistic regression analysis

bglmm: Multilevel data using Bayesian multilevel logistic regression analysis

References

Kavelaars X, Mulder J, Kaptein M (2020). “Decision-making with multiple correlated binary outcomes in clinical trials.” Statistical Methods in Medical Research, 29(11), 3265–3277. doi:10.1177/0962280220922256 .Kavelaars X, Mulder J, Kaptein M (2024). “Bayesian Multivariate Logistic Regression for Superiority and Inferiority Decision-Making under Observable Treatment Heterogeneity.” Multivariate Behavioral Research, 59(4), 859–882. doi:10.1080/00273171.2024.2337340 .Kavelaars X, Mulder J, Kaptein M (2023). “Bayesian multilevel multivariate logistic regression for superiority decision-making under observable treatment heterogeneity.” BMC Medical Research Methodology, 23(1). doi:10.1186/s12874-023-02034-z .

See also

Author

Maintainer: Xynthia Kavelaars xynthia.kavelaars@ou.nl (ORCID)

Other contributors:

Joris Mulder (ORCID) [thesis advisor]

Maurits Kaptein (ORCID) [thesis advisor]

Dutch Research Council (Grant no. 406.18.505) [funder]